Add Logout, Security and exception error handling

+++ spring-boot-2-todo-sample/src/main/java/com/in28minutes/springboot/web/controller/TodoController.java
@@ -4,31 +4,32 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 
 import javax.validation.Valid;
-import org.springframework.web.bind.WebDataBinder;
-import org.springframework.web.bind.annotation.InitBinder;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.propertyeditors.CustomDateEditor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
+import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.SessionAttributes;
 
 import com.in28minutes.springboot.web.model.Todo;
-import com.in28minutes.springboot.web.service.LoginService;
 import com.in28minutes.springboot.web.service.TodoService;
 
 @Controller
-@SessionAttributes("name")
 public class TodoController {
 
 	@Autowired
 	TodoService service;
 
 	@InitBinder
-	protected void initBinder(WebDataBinder binder) {
+	public void initBinder(WebDataBinder binder) {
+		// Date - dd/MM/yyyy
 		SimpleDateFormat dateFormat = new SimpleDateFormat("dd/MM/yyyy");
 		binder.registerCustomEditor(Date.class, new CustomDateEditor(
 				dateFormat, false));
@@ -36,20 +37,35 @@ public class TodoController {
 
 	@RequestMapping(value = "/list-todos", method = RequestMethod.GET)
 	public String showTodos(ModelMap model) {
-		String name = (String) model.get("name");
+		String name = getLoggedInUserName(model);
 		model.put("todos", service.retrieveTodos(name));
 		return "list-todos";
 	}
 
+	private String getLoggedInUserName(ModelMap model) {
+		Object principal = SecurityContextHolder.getContext()
+				.getAuthentication().getPrincipal();
+
+		if (principal instanceof UserDetails) {
+			return ((UserDetails) principal).getUsername();
+		}
+
+		return principal.toString();
+	}
+
 	@RequestMapping(value = "/add-todo", method = RequestMethod.GET)
 	public String showAddTodoPage(ModelMap model) {
-		model.addAttribute("todo", new Todo(0, (String) model.get("name"),
+		model.addAttribute("todo", new Todo(0, getLoggedInUserName(model),
 				"Default Desc", new Date(), false));
 		return "todo";
 	}
 
 	@RequestMapping(value = "/delete-todo", method = RequestMethod.GET)
 	public String deleteTodo(@RequestParam int id) {
+
+		if(id==1)
+			throw new RuntimeException("Something went wrong");
+
 		service.deleteTodo(id);
 		return "redirect:/list-todos";
 	}
@@ -62,19 +78,19 @@ public class TodoController {
 	}
 
 	@RequestMapping(value = "/update-todo", method = RequestMethod.POST)
-	public String updateTodo(ModelMap model, @Valid Todo todo, BindingResult result) {
+	public String updateTodo(ModelMap model, @Valid Todo todo,
+							 BindingResult result) {
 
 		if (result.hasErrors()) {
 			return "todo";
 		}
 
-		todo.setUser((String) model.get("name"));
+		todo.setUser(getLoggedInUserName(model));
 
 		service.updateTodo(todo);
 
 		return "redirect:/list-todos";
 	}
-
 
 	@RequestMapping(value = "/add-todo", method = RequestMethod.POST)
 	public String addTodo(ModelMap model, @Valid Todo todo, BindingResult result) {
@@ -83,7 +99,7 @@ public class TodoController {
 			return "todo";
 		}
 
-		service.addTodo((String) model.get("name"), todo.getDesc(), new Date(),
+		service.addTodo(getLoggedInUserName(model), todo.getDesc(), todo.getTargetDate(),
 				false);
 		return "redirect:/list-todos";
 	}
